src/frontend-module/requirements.txt
fastapi
uvicorn
psycopg2-binary
pika
jinja2
websockets